Add assertions to check validity of index
There are multiple command classes with a method that takes in indexes as arguments, without checking their validity.

Even though these indexes are guaranteed to be valid if everything is implemented correctly, it is better to use assertions in order to ensure and demonstrate the validity.

Let's put assertions, in every function that takes in indexes as arguments, to check their validity.